requirements.txt
aiogram
python-dotenv
tortoise-orm
asyncpg